+++ src/EventListener/Contao/GeneratePageListener.php
@@ -1,4 +1,4 @@
-<?php /** @noinspection PhpUndefinedNamespaceInspection, PhpUndefinedClassInspection */
+<?php
 
 namespace HeimrichHannot\PwaBundle\EventListener\Contao;
 
@@ -8,9 +8,6 @@ use Contao\PageModel;
 use Contao\PageRegular;
 use HeimrichHannot\PwaBundle\DataContainer\PageContainer;
 use HeimrichHannot\PwaBundle\Generator\ConfigurationFileGenerator;
-use HeimrichHannot\PwaBundle\HeaderTag\ManifestLinkTag;
-use HeimrichHannot\PwaBundle\HeaderTag\PwaHeadScriptTags;
-use HeimrichHannot\PwaBundle\HeaderTag\ThemeColorMetaTag;
 use HeimrichHannot\PwaBundle\Model\PwaConfigurationsModel;
 use HeimrichHannot\EncoreBundle\Asset\FrontendAsset;
 use Psr\Container\ContainerInterface;
@@ -21,10 +18,8 @@ readonly class GeneratePageListener implements ServiceSubscriberInterface
 {
     public function __construct(
         private ContainerInterface         $container,
-        private ManifestLinkTag            $manifestLinkTag,
-        private ThemeColorMetaTag          $colorMetaTag,
-        private PwaHeadScriptTags          $pwaHeadScriptTags,
         private ConfigurationFileGenerator $configurationGenerator
+
     ) {}
 
     public function __invoke(PageModel $pageModel, LayoutModel $layout, PageRegular $pageRegular): void
@@ -43,9 +38,6 @@ readonly class GeneratePageListener implements ServiceSubscriberInterface
             return;
         }
 
-        // $this->manifestLinkTag->setContent('/pwa/' . $rootPage->alias . '_manifest.json');
-        $this->colorMetaTag->setContent('#' . \ltrim($config->pwaThemeColor, '#'));
-
         if (!$jsonConfig = \json_encode($this->configurationGenerator->generateConfiguration($rootPage, $config), \JSON_UNESCAPED_UNICODE)) {
             return;
         }
@@ -57,7 +49,7 @@ readonly class GeneratePageListener implements ServiceSubscriberInterface
         <script>window.HuhPWA = JSON.parse('$jsonConfig')</script>
         HTML;
 
-        $GLOBALS['TL_HEAD']['something'] = $script;
+        $GLOBALS['TL_HEAD']['huh_pwa'] = $script;
 
         if ($this->container->has(FrontendAsset::class)) {
             $this->container->get(FrontendAsset::class)->addActiveEntrypoint('contao-pwa-bundle');